Allow utils to be used in fake
